Moved IsBucketNameValid to gslib/utils/text_utils.py
