+++ gslib/commands/iam.py
@@ -53,6 +53,7 @@ from gslib.utils.iam_helper import IsEqualBindings
 from gslib.utils.iam_helper import PatchBindings
 from gslib.utils.iam_helper import SerializeBindingsTuple
 from gslib.utils.retry_util import Retry
+from gslib.utils.text_util import IsBucketNameValid
 
 _SET_SYNOPSIS = """
   gsutil iam set [-afRr] [-e <etag>] file url ...
@@ -214,40 +215,6 @@ _DETAILED_HELP_TEXT = CreateHelpText(_SYNOPSIS, _DESCRIPTION)
 _get_help_text = CreateHelpText(_GET_SYNOPSIS, _GET_DESCRIPTION)
 _set_help_text = CreateHelpText(_SET_SYNOPSIS, _SET_DESCRIPTION)
 _ch_help_text = CreateHelpText(_CH_SYNOPSIS, _CH_DESCRIPTION)
-
-def _IsBucketNameValid(bucket_name):
-  """Check if string meets all bucket name requirements.
-
-  Requirements for bucket names defined at:
-  https://cloud.google.com/storage/docs/naming
-
-  Args:
-    Unicode tring, name of bucket. Full name, including provider prefix, i.e.:
-    'gs://my-bucket-name'
-  Returns:
-    Boolean, whether the bucket name is valid or not.
-  """
-
-  def _StripAllowedSymbols(s):
-    """Remove from a string the symbols '-', '_', and '.'"""
-    return ''.join((char for char in s if char not in ['-', '_', '.']))
-
-  if not '://' in bucket_name:
-    return False
-
-  prefix, url = bucket_name.split('://')
-
-  return all([
-    prefix.isalpha(),
-    prefix.islower(),
-    len(url) > 2,
-    len(url) < 64,
-    url[0].isalnum(),
-    url[-1].isalnum(),
-    not url.startswith('goog'),
-    not url.isupper(),
-    _StripAllowedSymbols(url).isalnum()
-  ])
 
 
 IAM_CH_CONDITIONS_WORKAROUND_MSG = (
@@ -513,7 +480,7 @@ class IamCommand(Command):
     # expecting to come across the -r, -f flags here.
     it = iter(self.args)
     for token in it:
-      if _IsBucketNameValid(token):
+      if IsBucketNameValid(token):
         patterns.append(token)
         break
       if token == '-d':
